feat: implement jsdocs for tasks and boards
